pyproject.toml
[project]
name = "cs431-rag"
version = "0.1.0"
description = "CS431 RAG"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerebras-cloud-sdk>=1.50.1",
    "fastapi>=0.119.0",
    "huggingface-hub[cli,hf-transfer]>=0.35.3",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "underthesea>=8.3.0",
    "sentencepiece>=0.2.1",
    "yt-dlp>=2025.11.12",
    "protobuf>=6.33.1",
    "faster-whisper>=1.2.1",
    "streamlit>=1.51.0",
    "rouge>=1.0.1",
    "pytest>=9.0.2",
    "langchain>=1.2.10",
    "pymilvus>=2.6.9",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "pytest-asyncio>=1.3.0",
    "langchain-google-genai>=4.2.1",
]


[tool.pytest]
testpaths = ["tests"]
